WIP: sort keys by UTF-16

+++ src/entry.rs
@@ -2,7 +2,7 @@ use std::collections::BTreeMap;
 
 #[derive(Clone, Debug)]
 pub struct Entry {
-  pub object: BTreeMap<Vec<u8>, Vec<u8>>,
+  pub object: BTreeMap<Vec<u16>, Vec<u8>>,
   pub next_key: Vec<u8>,
   pub next_val: Vec<u8>,
   pub complete: bool,
